rework func inc

+++ scripts/ubuntu/tear-down.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 
-SCRIPT_PATH=$(dirname "${BASH_SOURCE[0]}")
+SCRIPT_PATH=$(realpath "$(dirname "${BASH_SOURCE[0]}")/..")
+source ${SCRIPT_PATH}/functions.inc.sh
 
-source ${SCRIPT_PATH}/../functions.inc.sh
+conf ${1}
 
 if apt list --installed 2>/dev/null | grep -q "takserver";then
 	msg $warn "\nTAK Server clean up (takes a moment to stop TAK)"
@@ -15,4 +16,12 @@ if apt list --installed 2>/dev/null | grep -q "takserver";then
 	sudo -u postgres psql -c "DROP USER IF EXISTS martiuser;"
 fi
 
-rm -rf /opt/tak
+if [ -d "${RELEASE_PATH}" ]; then
+	${SCRIPT_PATH}/cert-bundler.sh ${TAK_ALIAS}
+
+	msg $danger "\nWiping ${RELEASE_PATH}"
+	rm -rf ${RELEASE_PATH}
+
+	msg $danger "\nWiping /opt/tak"
+	rm -rf /opt/tak
+fi
